refactor(docs): separate Markdown parsing and remove unused node state
